requirements.txt
streamlit
langchain
langchain-community
langchain-qdrant
langchain-ollama
langchain-huggingface
sentence-transformers
qdrant-client
pypdf
torchvision
ragas
datasets
pandas
langchain-google-genai
python-dotenv
